requirements.txt
mcp
slack_sdk
pydantic
python-dotenv
httpx
fastapi
uvicorn
sse-starlette
requests
supabase

